+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* Add an AuthorFollowerScorer
+* Rename Toot object realAccount() to author()
 
 ### v0.65.2
 * (Demo App) Don't show `OAUTH_ERROR_MSG` on network errors
@@ -8,7 +10,7 @@
 * Add cache summary to `getCurrentState()`'s storage info
 
 ## v0.65.0
-* Add `triggerMoarData()` method to TheAlgorithm
+* Add `triggerMoarData()` method to `TheAlgorithm`
 * Don't trigger unnecessarsy `maxCacheRecords` truncation calls
 * Explicitly mark `loadStartedAt` instead of relying on message passing to `setLoadingState()`
 * (Demo App) Add button/link to trigger pulling the next batch of historical user data to tune the algorithm